LayoutTests/imported/w3c/web-platform-tests/resize-observer/calculate-depth-for-node-expected.txt

CONSOLE MESSAGE: ResizeObserver loop completed with undelivered notifications.

Harness Error (FAIL), message = Script error.

FAIL "Calculate depth for node" algorithm with Shadow DOM assert_false: expected false got true
